+++ lib_i2s/api/i2s.h
@@ -105,15 +105,15 @@ typedef interface i2s_callback_if {
 #if defined(__XS2A__) || defined(__DOXYGEN__)
 
 /** Interface representing callback events that can occur during the
- *   operation of the I2S task. This is specific to the High Efficiency
+ *   operation of the I2S task. This is specific to the frame-based
  *   I2S master task which uses hardware generation of BCLK and transfers
  *   samples as arrays rather than individual channels, resulting in much
  *   less senesitivity to back pressure in the send/recieve cases.
  *   This interface is supported on xCORE200 processors only.
  */
-typedef interface i2s_he_callback_if {
+typedef interface i2s_frame_callback_if {
 
-  /**  I2S High Efficiency initialization event callback.
+  /**  I2S frame-based initialization event callback.
    *
    *   The I2S component will call this
    *   when it first initializes on first run of after a restart.
@@ -129,7 +129,7 @@ typedef interface i2s_he_callback_if {
    */
   void init(i2s_config_t &?i2s_config, tdm_config_t &?tdm_config);
 
-  /**  I2S High Efficiency restart check callback.
+  /**  I2S frame-based restart check callback.
    *
    *   This callback is called once per frame. The application must return the
    *   required restart behaviour.
@@ -143,7 +143,7 @@ typedef interface i2s_he_callback_if {
   /**  Receive an incoming sample.
    *
    *   This callback will be called when a new sample is read in by the I2S
-   *   High Efficiency component.
+   *   frame-based component.
    *
    *  \param num_out    The number of input channels contained within the array.
    *  \param sample     The sample data array as signed 32-bit values.  The component
@@ -155,7 +155,7 @@ typedef interface i2s_he_callback_if {
 
   /** Request an outgoing sample.
    *
-   *  This callback will be called when the I2S High Efficiency component needs
+   *  This callback will be called when the I2S frame-based component needs
    *  a new sample.
    *
    *  \param num_out    The number of output channels contained within the array.
@@ -166,7 +166,7 @@ typedef interface i2s_he_callback_if {
    */
   void send(size_t num_out, int32_t sample[num_out]);
 
-} i2s_he_callback_if;
+} i2s_frame_callback_if;
 
 #endif // __XS2A__
 
@@ -205,16 +205,16 @@ void i2s_master(client i2s_callback_if i2s_i,
 
 #if defined(__XS2A__) || defined(__DOXYGEN__)
 
-/** I2S High Efficiency master component **for xCORE200 only**
+/** I2S frame-based master component **for xCORE200 only**
  *
  *  This task performs I2S on the provided pins. It will perform callbacks over
- *  the i2s_callback_if interface to get/receive data from the application
- *  using this component.
+ *  the i2s_frame_callback_if interface to get/receive frames of data from the
+ *  application using this component.
  *
  *  The component performs I2S master so will drive the word clock and
  *  bit clock lines.
  *
- *  \param i2s_i          The I2S  High Efficiency callback interface to connect to
+ *  \param i2s_i          The I2S frame callback interface to connect to
  *                        the application
  *  \param p_dout         An array of data output ports
  *  \param num_out        The number of output data ports
@@ -226,7 +226,7 @@ void i2s_master(client i2s_callback_if i2s_i,
  *  \param bclk           A clock that will get configured for use with
  *                        the bit clock
  */
-void i2s_he_master(client i2s_he_callback_if i2s_i,
+void i2s_frame_master(client i2s_frame_callback_if i2s_i,
                 out buffered port:32 (&?p_dout)[num_out],
                 static const size_t num_out,
                 in buffered port:32 (&?p_din)[num_in],
@@ -345,7 +345,7 @@ void i2s_tdm_master(client interface i2s_callback_if tdm_i,
         clock mclk);
 
 #include <i2s_master_impl.h>
-#include <i2s_master_he_impl.h>
+#include <i2s_frame_master_impl.h>
 #include <i2s_slave_impl.h>
 #include <tdm_master_impl.h>
 #include <i2s_tdm_master_impl.h>